# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication and Authorization API",
        "title": "Auth API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/pdf/pdf-to-jpg": {
            "post": {
                "description": "Convert a PDF file's pages into JPG format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-to-jpg"
                ],
                "summary": "Convert PDF to JPG",
                "parameters": [
                    {
                        "description": "PDF file ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PDFToJPGRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/pdf-to-jpg/{id}": {
            "get": {
                "description": "Retrieve the status of the conversion job by Job ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-to-jpg"
                ],
                "summary": "Get PDF to JPG conversion job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PDFToJPGJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/pdf-to-word": {
            "post": {
                "description": "PDF faylni Word formatiga o‘tkazish",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF to Word"
                ],
                "summary": "Convert PDF to Word",
                "parameters": [
                    {
                        "description": "PDF to Word request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PDFToWordRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/pdf-to-word/{id}": {
            "get": {
                "description": "PDF to Word ish holatini ko‘rish",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF to Word"
                ],
                "summary": "Get PDF to Word Job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PDFToWordJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/ppt-to-pdf": {
            "post": {
                "description": "PowerPoint faylni PDF formatga o‘tkazish",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PowerPoint to PDF"
                ],
                "summary": "Convert PowerPoint to PDF",
                "parameters": [
                    {
                        "description": "PowerPoint to PDF request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PowerPointToPDFRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "job_id qaytadi",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/ppt-to-pdf/{id}": {
            "get": {
                "description": "Konvertatsiya holatini olish",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PowerPoint to PDF"
                ],
                "summary": "Get PowerPoint→PDF Job Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PowerPointToPDFJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/protect": {
            "post": {
                "description": "PDF fayliga parol qo‘shadi (foydalanuvchi tomonidan berilgan parol asosida)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Security"
                ],
                "summary": "Protect PDF with password",
                "parameters": [
                    {
                        "description": "Protect request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProtectPDFRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/pdf/protect/{id}": {
            "get": {
                "description": "Parollangan PDF faylga oid ish holatini qaytaradi",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Security"
                ],
                "summary": "Get protected PDF job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProtectPDFJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.PDFToJPGJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Job creation timestamp",
                    "type": "string"
                },
                "id": {
                    "description": "Job ID",
                    "type": "string"
                },
                "input_file_id": {
                    "description": "Input file ID",
                    "type": "string"
                },
                "output_file_ids": {
                    "description": "List of generated JPG file IDs",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "description": "Status of the job (pending, processing, done, failed)",
                    "type": "string"
                },
                "user_id": {
                    "description": "User ID (nullable)",
                    "type": "string"
                },
                "zip_file_id": {
                    "description": "ZIP file ID (nullable, can be nil)",
                    "type": "string"
                }
            }
        },
        "models.PDFToJPGRequest": {
            "type": "object",
            "required": [
                "input_file_id"
            ],
            "properties": {
                "input_file_id": {
                    "description": "Required input file ID",
                    "type": "string"
                }
            }
        },
        "models.PDFToWordJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.PDFToWordRequest": {
            "type": "object",
            "required": [
                "input_file_id"
            ],
            "properties": {
                "input_file_id": {
                    "type": "string"
                }
            }
        },
...
        "models.PowerPointToPDFJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.PowerPointToPDFRequest": {
            "type": "object",
            "required": [
                "input_file_id"
            ],
            "properties": {
                "input_file_id": {
                    "type": "string"
                }
            }
        },
        "models.ProtectPDFJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "description": "← *pointer (✔ to‘g‘ri)",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.ProtectPDFRequest": {
            "type": "object",
            "required": [
                "input_file_id",
                "password"
            ],
            "properties": {
                "input_file_id": {
                    "description": "PDF fayl ID",
                    "type": "string"
                },
                "password": {
                    "description": "Qo‘yiladigan parol",
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}